history has been added ready for tag. There are some issues:
New issue
Summary: Class LP linear programming is not in a usable state. Needs to be fixed.
Labels: Priority-High
This needs to be fixed before the version number will be bumped up to 0.14. To fix this
the cvxopt.solvers.lp functionality will have to be investigated.

New issue
Summary: Missing unittests for PolicyIterationModified
Labels: Priority-High
These unit tests should be implemented as soon as possible. Will not make first release without them.