QLearning unittest fix and LP verbosity problem
New issue
Summary: It is not possible to switch verbose to False in the LP class
Status: Started
Labels: Type-Defect Priority-Medium Usability
The LP class uses the module cvxopt for linear programming functionality. When
LP.verbose is set to False then the cvxopt.solvers.options['show_progress']
should also be False. Need to find a way to integrate the options component into
the LP class, or find another way around it.