[tests] Remove unused travis deps
The CVXOPT python package has some optional dependencies that must be
altered in the setup.py file for support to be compiled in. GLPK is one
if these. At the moment the setup.py script is not being altered while
installing dependencies, so downloading cvxopts extra dependencies is
useless.

Move the test command from .travis.yml to runtests.sh to allow easy
testing on workstations.